Refactor the github mondo refresh system
The principles are simple:

every file in make should depend on other files, not internal make goals.

The mondo.owl and mondo.sssom.tsv targets simply depend on the cloned repo.

The cloned repo logic was flipped. The main dependency (tmp/mondo_repo_built) which is always run (.FORCE) controls wether the costly "clone rebuild" phony goal is executed.

If nothing is executed the, file representing the goal (tmp/mondo_repo_built) is not updated which means nothing underneath it will be updated.